backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.0
drizzle-orm==0.30.1
asyncpg==0.29.0
alembic==1.13.1
httpx==0.26.0
python-multipart==0.0.9